settings: Disabling set 'id' to nobody
This patch temporarily disable set 'id' to nobody. Storage fails to
write in nRF24 keys file since nrfd does not have permission to write
in this file. In order to allow storage to write on nRF24 keys file a
temporarily approach is to disable set nobody while studying a better
one.